app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
celery==5.4.0
redis==5.0.8
pika==1.3.2
httpx==0.27.2
pytest==8.3.3
flake8==7.1.1
